Final stage
In this stage, I have completed the program. The user is now able to retrieve an accurate t value. 

The validity of this program is confirmed a university approved t test calculator from the faculty of science and engineering. 

Both the uni approved calculator and mine produce the exact same result... YAY!